회원가입 ID중복확인 & 오류 예외처리 & MVC 구성 다듬기
