+++ .gitignore
@@ -1,7 +1,7 @@
 # See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
 
 #database 보안상 gitignore 해야함! pw유출 조심!
-/database.json
+/config/database.json
 
 # dependencies
 /node_modules
