--- database.json
@@ -1,7 +1,0 @@
-{
-    "host": "DB서버 엔드포인트", 
-    "user": "사용자이름",     
-    "password": "비밀번호", 
-    "port": "3306", 
-    "database": "database 이름" 
-}
